Fix warnings.

+++ RFPerformance.m
@@ -64,16 +64,16 @@
 	mach_msg_type_number_t infoCount = HOST_VM_INFO_COUNT;
     kern_return_t kernReturn = host_statistics(mach_host_self(), HOST_VM_INFO, (host_info_t)&vmStats, &infoCount);
 	
-    if (kernReturn == KERN_SUCCESS){
-        NSLog(@"free: %u\nactive: %u\ninactive: %u\nwire: %u\nzero fill: %u\nreactivations: %u\npageins: %u\npageouts: %u\nfaults: %u\ncow_faults: %u\nlookups: %u\nhits: %u",
-			  vmStats.free_count * vm_page_size,
-			  vmStats.active_count * vm_page_size,
-			  vmStats.inactive_count * vm_page_size,
-			  vmStats.wire_count * vm_page_size,
-			  vmStats.zero_fill_count * vm_page_size,
-			  vmStats.reactivations * vm_page_size,
-			  vmStats.pageins * vm_page_size,
-			  vmStats.pageouts * vm_page_size,
+    if (kernReturn == KERN_SUCCESS) {
+        NSLog(@"free: %lu\nactive: %lu\ninactive: %lu\nwire: %lu\nzero fill: %lu\nreactivations: %lu\npageins: %lu\npageouts: %lu\nfaults: %u\ncow_faults: %u\nlookups: %u\nhits: %u",
+			  (unsigned long)vmStats.free_count * vm_page_size,
+			  (unsigned long)vmStats.active_count * vm_page_size,
+			  (unsigned long)vmStats.inactive_count * vm_page_size,
+			  (unsigned long)vmStats.wire_count * vm_page_size,
+			  (unsigned long)vmStats.zero_fill_count * vm_page_size,
+			  (unsigned long)vmStats.reactivations * vm_page_size,
+			  (unsigned long)vmStats.pageins * vm_page_size,
+			  (unsigned long)vmStats.pageouts * vm_page_size,
 			  vmStats.faults,
 			  vmStats.cow_faults,
 			  vmStats.lookups,
